Add notification processing and notification kind test
