+++ kifi-backend/modules/eliza/app/com/keepit/eliza/model/NotificationRepo.scala
@@ -6,10 +6,13 @@ import com.keepit.common.db.slick.DBSession.RSession
 import com.keepit.common.db.slick.{ DataBaseComponent, DbRepo, Repo }
 import com.keepit.common.time.Clock
 import com.keepit.model.User
+import com.keepit.notify.model.NKind
 import org.joda.time.DateTime
 
 @ImplementedBy(classOf[NotificationRepoImpl])
 trait NotificationRepo extends Repo[Notification] {
+
+  def getLastByUserAndKind(userId: Id[User], kind: NKind)(implicit session: RSession): Option[Notification]
 
 }
 
@@ -37,5 +40,11 @@ class NotificationRepoImpl @Inject() (
   def deleteCache(model: Notification)(implicit session: RSession): Unit = {}
 
   def invalidateCache(model: Notification)(implicit session: RSession): Unit = {}
+
+  def getLastByUserAndKind(userId: Id[User], kind: NKind)(implicit session: RSession): Option[Notification] = {
+    val kindStr = kind.name
+    val query = (for (row <- rows if row.userId === userId && row.kind === kindStr) yield row).sortBy(_.createdAt.desc)
+    query.firstOption
+  }
 
 }
